% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1.R
\name{clipLevel1}
\alias{clipLevel1}
\title{Clip LVIS Level1 data}
\usage{
clipLevel1(level1_waveform, output, xleft, xright, ybottom, ytop)
}
\arguments{
\item{level1_waveform}{h5file; S4 object of class H5File}

\item{output}{path where to save the new h5file}

\item{xleft}{numeric. left x coordinates of rectangles.}

\item{xright}{numeric. right x coordinates of rectangles.}

\item{ybottom}{numeric. bottom y coordinates of rectangles.}

\item{ytop}{numeric. top y coordinates of rectangles.}
}
\value{
Returns An object of class H5File; subset of LVIS Level1 data
}
\description{
Clip LVIS Level1 data within a given bounding coordinates
}
\examples{

#' LVIS level 2 file path
level1_filepath = system.file("extdata", "LVIS_Mondah_level1b.SPLIT.8.h5", package="rLVIS")

# Rectangle
xleft = 9.35986
xright = 9.35988
ybottom = 0.5786
ytop = 0.5790

#' Reading LVIS level 2 file
level1_waveform = readLevel1b(level1_filepath)

output = tempfile(fileext="h5")

clipped_waveform = clipLevel1(level1_waveform, output, xleft, xright, ybottom, ytop)

}
\author{
Caio Hamamura
}
