% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1.R
\name{clipLevel1Geometry}
\alias{clipLevel1Geometry}
\title{Clip LVIS Level1 data by geometry}
\usage{
clipLevel1Geometry(level1_waveform, output, polygon_spdf)
}
\arguments{
\item{level1_waveform}{h5file; S4 object of class H5File}

\item{output}{path where to save the new h5file}

\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}
}
\value{
Returns An object of class H5File; subset of LVIS Level1 data
}
\description{
Clip LVIS Level1 data within a given bounding coordinates
}
\examples{

#' LVIS level 2 file path
level1_filepath = system.file("extdata", "LVIS_Mondah_level1b.SPLIT.8.h5", package="rLVIS")

#' Reading LVIS level 2 file
level1_waveform = readLevel1b(level1_filepath)

# Polgons file path
polygons_filepath <- system.file("extdata", "lvis_level1_polygons.shp", package="rLVIS")

# Reading LVIS level 2 file
polygon_spdf<-raster::shapefile(polygons_filepath)

output = tempfile(fileext="h5")

clipped_waveform = clipLevel1Geometry(level1_waveform, output, polygon_spdf)

}
\author{
Caio Hamamura
}
