% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridMetrics.R
\name{GridMetrics}
\alias{GridMetrics}
\title{Compute a Series of Grid Metrics}
\usage{
GridMetrics(level2_spdf, func, res)
}
\arguments{
\item{level2_spdf}{LVIS l2 dataset; object of class \code{SpatialPointsDataFrame}}

\item{func}{the function to be applied to each cell}

\item{res}{spatial resolution for the output raster layer}
}
\value{
Returns raster layer (s) of selected LVIS metric (s)
}
\description{
This function computes a series of user-defined descriptive statistics for a LVIS dataset within
each grid cell
}
\examples{
\dontrun{
#' LVIS level 2 file path
level2_filepath_zip <- system.file("extdata", "LVIS_Mondah_level2.zip", package="rLVIS")
unzip(level2_filepath_zip, exdir = tempdir())
level2_filepath <- file.path(tempdir(), "LVIS_Mondah_level2.txt")

#' Reading LVIS level 2 file
level2_spdf<-readLevel2(level2path=level2_filepath,spdf=TRUE,glatlon=TRUE)

#' Define your own function
mySetOfMetrics = function(x)
{
metrics = list(
   min =min(x), # Min of z
   max = max(x), # Max of z
   mean = mean(x), # Mean of z
   sd = sd(x)# Sd of z
 )
 return(metrics)
}

#' Computing serie of LVIS metrics
mlvis<-GridMetrics(level2_spdf=level2_spdf,mySetOfMetrics(ZT), res=0.0005)

#' Computing single LVIS metrics
maxRH100<-GridMetrics(level2_spdf=level2_spdf,max(RH100), res=0.0005)

#' Computing single LVIS metrics
ZGmean<-GridMetrics(level2_spdf=level2_spdf,mean(ZG), res=0.0005)
rasterVis::plot3D(ZGmean, col="gray")
}
}
\author{
Carlos Alberto Silva (This function has been adapted from the grid_metrics function in lidR package. All credits to Roussel et al. 2019).
}
