% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotLevel2}
\alias{plotLevel2}
\title{LVIS Level2 data visualization in 2-D}
\usage{
plotLevel2(level2_spdf, color, colorPalette,...)
}
\arguments{
\item{level2_spdf}{LVIS l2 dataset; object of class \code{SpatialPointsDataFrame}}

\item{color}{The field name used to color the points.  Default is RH100}

\item{colorPalette}{A vector defining the color  palette.}

\item{...}{passing arguments on to the plot function}
}
\value{
A 2-D figure of LVIS Level2 data;
}
\description{
This function plots LVIS Level2 in 2-D
}
\examples{

# LVIS level2 file path
level2_filepath_zip <- system.file("extdata", "LVIS_Mondah_level2.zip", package="rLVIS")
unzip(level2_filepath_zip, exdir = tempdir())
level2_filepath <- file.path(tempdir(), "LVIS_Mondah_level1b.txt")

# Reading LVIS level1b file
level2_spdf<-readLevel2(level2path=level2_filepath, spdf=TRUE, glatlon=TRUE)

#' Plot LVIS Level2 data
head(level2_spdf@data)
#plotLevel2(level2_spdf=level2_spdf, color = "RH100",
#           colorPalette = c("blue","green","yellow","red"),
#           axes=T)
grid()
}
\author{
Carlos Alberto Silva.
}
