% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{Scatterplot of a 1:1 comparison}
\usage{
plotStats(x, y, n, colours = c("blue", "green", "yellow", "red"),
  xlim = NULL, ylim = NULL, colfill = "gray", xlab = NULL,
  ylab = NULL, title = "GEDI vs ALS", legend.position = c(0.8, 0.2),
  stats.position = c(50, 400, 5), base_size = 16, stat.size = 3.5,
  legend.size = c(8, 8, 3, 1), fit.line.col = c("black", "red"),
  x_axis = TRUE, y_axis = TRUE)
}
\arguments{
\item{x}{h5file; S4 object of class H5File}

\item{y}{dataframe containing LVIS level 2 data}

\item{n}{size}

\item{colours}{h5file; S4 object of class H5File}

\item{xlim}{dataframe containing LVIS level 2 data}

\item{ylim}{h5file; S4 object of class H5File}

\item{colfill}{dataframe containing LVIS level 2 data}

\item{xlab}{h5file; S4 object of class H5File}

\item{ylab}{dataframe containing LVIS level 2 data}

\item{title}{h5file; S4 object of class H5File}

\item{legend.position}{dataframe containing LVIS level 2 data}

\item{stats.position}{dataframe containing LVIS level 2 data}

\item{base_size}{dataframe containing LVIS level 2 data}

\item{stat.size}{dataframe containing LVIS level 2 data}

\item{legend.size}{dataframe containing LVIS level 2 data}

\item{fit.line.col}{dataframe containing LVIS level 2 data}

\item{x_axis}{dataframe containing LVIS level 2 data}

\item{y_axis}{dataframe containing LVIS level 2 data}
}
\value{
Returns A list containing statistical parameters and the Scatterplot of the 1:1 comparison
}
\description{
Compute and plot stastical parameters of 1:1 comparison
}
\examples{
# Importing libraries
library(raster)
library(rasterVis)
library(viridis)
library(gridExtra)

# Importing dataset
sf_agb1ha_path <- system.file("extdata", "sf_agb_1ha.tif", package="rLVIS")
lf_agb1ha_path <- system.file("extdata", "lf_agb_1ha.tif", package="rLVIS")

sf_agb<-raster(sf_agb1ha_path)
lf_agb<-raster(lf_agb1ha_path)

# Ploting AGB maps
s <- stack(sf_agb,lf_agb)
agb.maps<-levelplot(s,
         layout=c(1, 2),
         margin=FALSE,
         colorkey=list(
           space='right',
           labels=list(at=seq(0, 500, 50), font=4),
           axis.line=list(col='black'),
           width=1
         ),
         par.settings=list(
           strip.border=list(col='transparent'),
           strip.background=list(col='transparent'),
           axis.line=list(col='transparent')
         ),
         scales=list(draw=TRUE),
         col.regions=viridis,
         at=seq(0, 500, len=101),
         names.attr=c("SF_AGB","LF_AGB"))

# Ploting Stats
colours<-viridis(10)
base_size=15
legend.position= c(0.85, 0.3)
stat.size=5
stats.position=c(100,400,50)
base_size=base_size
xlim=c(0,500)
legend.size=c(8,15,10,2)
ylim=c(0,500)
ylab="LF-derived AGB (Mg/ha)"
xlab="SF-derived AGB (Mg/ha)"
fit.line.col=c("black","gray")
title="SF vs LF lidar"
x_axis=TRUE
y_axis=TRUE

\dontrun{
x11()
}
agb.comp<-plotStats(y=getValues(lf_agb),
                 x=getValues(sf_agb),
                 colours=colours,
                 legend.position= legend.position,
                 stat.size=stat.size,
                 stats.position=stats.position,
                 base_size=base_size,
                 xlim=xlim,
                 legend.size=legend.size,
                 ylim=ylim,
                 ylab=ylab,
                 xlab=xlab,
                 fit.line.col=fit.line.col,
                 title=title,
                 x_axis=x_axis,
                 y_axis=y_axis)

# Combining plots
grid.arrange(agb.maps,agb.comp$plotg, nrow = 1)

}
\author{
Carlos Alberto Silva.
}
