% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotWaveform}
\alias{plotWaveform}
\title{Display LVIS Waveform}
\usage{
plotWaveform(level1b,level2,shotnum,plot2=TRUE,...)
}
\arguments{
\item{level1b}{h5file; S4 object of class H5File}

\item{level2}{dataframe containing LVIS level 2 data}

\item{shotnum}{LVIS shot number to display}

\item{plot2}{if TRUE, plot both Level1b and Level2. If FALSE, plot only Level1b data}

\item{...}{passing arguments on to the plot function}
}
\value{
Returns 2-D scatterplot of the LVIS waveform
}
\description{
This function plots LVIS level1b and level 2 data
}
\examples{

# LVIS level1b file path
level1b_filepath_zip <- system.file("extdata", "LVIS_Mondah_level1b.zip", package="rLVIS")
unzip(level1b_filepath_zip, exdir = tempdir())
level1b_filepath <- file.path(tempdir(), "LVIS_Mondah_level1b.h5")

# Reading LVIS level1b file
level1b<-readLevel1b(level1bpath=level1b_filepath)

# LVIS level2 file path
level2_filepath_zip <- system.file("extdata", "LVIS_Mondah_level2.zip", package="rLVIS")
unzip(level2_filepath_zip, exdir = tempdir())
level2_filepath <- file.path(tempdir(), "LVIS_Mondah_level1b.txt")

# Reading LVIS level1b file
level2_spdf<-readLevel2(level2path=level2_filepath, spdf=TRUE, glatlon=TRUE)

#'Plotting LVIS waveforms
plotWaveform(level1b=level1b,level2=level2_spdf,
             shotnum=10964985,plot2=TRUE,xlab="Relative amplitude (\%)", ylab="Height (m)")

}
\author{
Carlos Alberto Silva.
}
