% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_processExtractMetrics.R
\name{processMetricsHDFlvis}
\alias{processMetricsHDFlvis}
\title{Process and calculate metrics for HDF LVIS level1}
\usage{
processMetricsHDFlvis(input, level2 = character(0),
  bounds = numeric(0), rhRes = 5, laiRes = 10, laiH = 30,
  noRHgauss = FALSE, gTol = 0, fhdHistRes = 0.001,
  forcePsigma = FALSE, bayesGround = FALSE, noRoundCoord = FALSE,
  dcBias = numeric(0), nSig = 0, seed = 1L, hNoise = 0,
  linkNoise = numeric(0), linkFsig = 5.5, linkPsig = 0.764331,
  trueSig = 5, bitRate = 12L, maxDN = 4096, renoise = FALSE,
  newPsig = 1, oldPsig = 0.764331, addDrift = 0, missGround = 0L,
  minGap = numeric(0), meanN = 0, thresh = 1e-08, varNoise = TRUE,
  varScale = 3, statsLen = 10, noiseTrack = FALSE, sWidth = 0.3,
  psWidth = 0, msWidth = 0, preMatchF = FALSE, postMatchF = FALSE,
  pFile = character(0), gWidth = 1.2, minGsig = 0.764331,
  minWidth = 3, medNoise = FALSE, varDrift = FALSE,
  driftFac = numeric(0), pSigma = 1, gold = FALSE,
  deconTol = 1e-07)
}
\arguments{
\item{input}{waveform  input filename}

\item{level2}{level2 filename for LVIS ZG}

\item{bounds}{only analyse data within bounds.}

\item{rhRes}{percentage energy resolution of RH metrics}

\item{laiRes}{LAI profile resolution in metres}

\item{laiH}{height to calculate LAI to}

\item{noRHgauss}{do not fit Gaussians}

\item{gTol}{ALS ground tolerance. Used to calculate slope.}

\item{fhdHistRes}{waveform intensity resolution to use when calculating FHD from histograms}

\item{forcePsigma}{do not read pulse sigma from file}

\item{bayesGround}{use Bayseian ground finding}

\item{noRoundCoord}{do not round up coords when outputting}

\item{dcBias}{mean noise level}

\item{nSig}{noise sigma}

\item{seed}{random number seed}

\item{hNoise}{hard threshold noise as a fraction of integral}

\item{linkNoise}{apply Gaussian noise based on link margin at a cover}

\item{linkFsig}{footprint width to use when calculating and applying signal noise}

\item{linkPsig}{pulse width to use when calculating and applying signal noise}

\item{trueSig}{true sigma of background noise}

\item{bitRate}{digitisation bit rate}

\item{maxDN}{maximum DN}

\item{renoise}{remove noise from truth before applying new noise level}

\item{newPsig}{new value for pulse width, when lengthening pulse}

\item{oldPsig}{old value for pulse width if not defined in waveform file, when lengthening pulse}

\item{addDrift}{apply detector background drift}

\item{missGround}{assume ground is missed to assess RH metrics}

\item{minGap}{delete signal beneath min detectable gap fraction}

\item{meanN}{mean noise level, if using a predefined mean level}

\item{thresh}{noise threshold, if using a predefined noise threshold}

\item{varNoise}{use a variable noise threshold}

\item{varScale}{variable noise threshold scale (multiple of stdev above mean to set threshold)}

\item{statsLen}{length to calculate noise stats over for varNoise}

\item{noiseTrack}{use noise tracking}

\item{sWidth}{smoothing width, after densoising}

\item{psWidth}{smoothing width, before denoising}

\item{msWidth}{smoothing width, after noise stats, before denoising}

\item{preMatchF}{matched filter before denoising}

\item{postMatchF}{matched filter after denoising}

\item{pFile}{read pulse file, for deconvolution and matched filters}

\item{gWidth}{Gaussian parameter selection smoothing width}

\item{minGsig}{minimum Gaussian sigma to fit}

\item{minWidth}{minimum feature width in bins}

\item{medNoise}{use median stats rather than mean}

\item{varDrift}{correct detector drift with variable factor}

\item{driftFac}{fix drift with constant drift factor}

\item{pSigma}{pulse width to smooth by if using Gaussian pulse}

\item{gold}{deconvolve with Gold's method}

\item{deconTol}{deconvolution tolerance}
}
\description{
Process and calculate metrics for HDF LVIS level1
}
