% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readLevel1b}
\alias{readLevel1b}
\title{Read LVIS Level1b data}
\usage{
readLevel1b(level1bpath)
}
\arguments{
\item{level1bpath}{file path pointing to LVIS level1b data (H5 format)}
}
\value{
H5File; S4 object of class H5File;
}
\description{
This function reads LVIS level1b data
}
\examples{

# LVIS level1b file path
level1b_filepath_zip <- system.file("extdata", "LVIS_Mondah_level1b.zip", package="rLVIS")
unzip(level1b_filepath_zip, exdir = tempdir())
level1b_filepath <- file.path(tempdir(), "LVIS_Mondah_level1b.h5")

# Reading LVIS level1b file
level1b<-readLevel1b(level1bpath=level1b_filepath)


}
\seealso{
\code{\link[h5]{h5file}} in the \emph{h5} package.
}
\author{
Carlos Alberto Silva. This function calls \emph{h5file} function from h5 package (Author: Mario Annau)
}
