% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel2.R
\name{readLevel2}
\alias{readLevel2}
\title{Read LVIS Level2 data}
\usage{
readLevel2(level2path, spdf=TRUE, glatlon=TRUE)
}
\arguments{
\item{level2path}{file path pointing to LVIS level2 data (txt format)}

\item{spdf}{if true, outups an object of class \code{SpatialPointsDataFrame}}

\item{glatlon}{if true, GLON and GLAT will be used for creating the \code{SpatialPointsDataFrame}.
If false, TLON and TLAT will be used instead}
}
\value{
An object of class \code{SpatialPointsDataFrame} or \code{data.table};
}
\description{
This function reads LVIS level2 data
}
\examples{

# LVIS level2 file path
level2_filepath_zip <- system.file("extdata", "LVIS_Mondah_level2.zip", package="rLVIS")
unzip(level2_filepath_zip, exdir = tempdir())
level2_filepath <- file.path(tempdir(), "LVIS_Mondah_level2.txt")

# Reading LVIS level1b file
level2_spdf<-readLevel2(level2path=level2_filepath, spdf=TRUE, glatlon=TRUE)

#' Plot LVIS Level2 data
plotLevel2(level2_spdf=level2_spdf, color = "RH100",
           colorPalette = c("blue","green","yellow","red"))


}
\author{
Carlos Alberto Silva.
}
